pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reaperbot"
version = "0.1.0"
description = "Discord voice relay bot for REAPER/ReaStream audio workflows"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ReaperBot Contributors" }]
dependencies = [
  "discord.py[voice]>=2.4.0",
  "python-dotenv>=1.0.0",
  "PyNaCl>=1.5.0"
]
keywords = ["discord", "audio", "reaper", "voice", "bot", "raspberry-pi"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Topic :: Multimedia :: Sound/Audio",
  "Topic :: Communications :: Chat"
]

[project.scripts]
reaperbot = "main:main"
